requirements.txt
asgiref==3.3.1
boto==2.49.0
dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.6.0
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.0.4
psycopg2==2.8.6
python-decouple==3.4
python-http-client==3.3.1
pytz==2020.5
sendgrid==6.4.8
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
whitenoise==5.2.0
